Moved shift tile functionality to shift_tile
