+++ backend/src/test/tile/test_tile_geometry.py
@@ -189,23 +189,6 @@ def test_center_on_entrance_b_causes_entrance_to_have_coordinate_0_0(tile_geomet
     assert actual.get_y() == 0
 
 
-def test_shift_tile_0_0_to_1_2_returns_1_2(tile_geometry_a):
-    tile_to_move = Tile(0, 0, None)
-    tile_to_shift_to = Tile(1, 2, None)
-    actual = tile_geometry_a.shift_tile(tile_to_move, tile_to_shift_to)
-    assert actual == tile_to_shift_to
-
-
-def test_shift_tile_list_0_0_and_0_1_to_1_0_returns_1_0_and_1_1(tile_geometry_a):
-    tiles_to_move = [Tile(0, 0, None), Tile(0, 1, None)]
-    tile_to_shift_to = Tile(1, 0, None)
-
-    actual = tile_geometry_a.shift_tile_list(tiles_to_move, tile_to_shift_to)
-
-    expected = [Tile(1, 0, None), Tile(1, 1, None)]
-    assert actual == expected
-
-
 def test_center_tunnel_on_top_of_burrow_on_waypoint_a_causes_waypoints_to_match_coordinates(tile_geometry_b,
                                                                                             waypoint_pojo_b):
     room_one = Burrow()
